Add files via upload
Unsupervised Machine Learning - Clustering using K-Means
K-Means clustering is an unsupervised learning technique that lacks predefined categories or groups. The primary objectives of this algorithm is to group similar data points together, with the number of groups determined by the value of K. The algorithm follows an iterative approach and assigns each data point to one of the K groups based on the features that are provided.